+++ src/Encoder.php
@@ -1,8 +1,4 @@
 <?php
-/**
- * @package axy\codecs\base64vlq
- * @author Oleg Grigoriev <go.vasac@gmail.com>
- */
 
 namespace axy\codecs\base64vlq;