+++ tests/BaseTestCase.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace axy\codecs\base64vlq\tests;
+
+use PHPUnit\Framework\TestCase;
+
+abstract class BaseTestCase extends TestCase
+{
+}