+++ tests/SignedTest.php
@@ -1,8 +1,4 @@
 <?php
-/**
- * @package axy\codecs\base64vlq
- * @author Oleg Grigoriev <go.vasac@gmail.com>
- */
 
 namespace axy\codecs\base64vlq\tests;
 
@@ -11,25 +7,20 @@ use axy\codecs\base64vlq\Signed;
 /**
  * coversDefaultClass axy\codecs\base64vlq\Signed
  */
-class SignedTest extends \PHPUnit_Framework_TestCase
+class SignedTest extends BaseTestCase
 {
     /**
      * covers ::encode
      * covers ::decode
      * @dataProvider providerEncodeDecode
-     * @param int $signed
-     * @param int $encoded
      */
-    public function testEncodeDecode($signed, $encoded)
+    public function testEncodeDecode(int $signed, int $encoded)
     {
         $this->assertSame($encoded, Signed::encode($signed));
         $this->assertSame($signed, Signed::decode($encoded));
     }
 
-    /**
-     * @return array
-     */
-    public function providerEncodeDecode()
+    public static function providerEncodeDecode(): array
     {
         return [
             [0, 0],
@@ -48,16 +39,13 @@ class SignedTest extends \PHPUnit_Framework_TestCase
      * @param int[] $sBlock
      * @param int[] $eBlock
      */
-    public function testEncodeDecodeBlock($sBlock, $eBlock)
+    public function testEncodeDecodeBlock(array $sBlock, array $eBlock)
     {
         $this->assertSame($eBlock, Signed::encodeBlock($sBlock));
         $this->assertSame($sBlock, Signed::decodeBlock($eBlock));
     }
 
-    /**
-     * @return array
-     */
-    public function providerEncodeDecodeBlock()
+    public static function providerEncodeDecodeBlock(): array
     {
         return [
             [[1, -234, 0, 134], [2, 469, 0, 268]],