+++ tests/errors/InvalidVLQSequenceTest.php
@@ -1,17 +1,14 @@
 <?php
-/**
- * @package axy\codecs\base64vlq
- * @author Oleg Grigoriev <go.vasac@gmail.com>
- */
 
 namespace axy\codecs\base64vlq\tests\errors;
 
 use axy\codecs\base64vlq\errors\InvalidVLQSequence;
+use axy\codecs\base64vlq\tests\BaseTestCase;
 
 /**
  * coversDefaultClass axy\codecs\base64vlq\errors\InvalidVLQSequence
  */
-class InvalidVLQSequenceTest extends \PHPUnit_Framework_TestCase
+class InvalidVLQSequenceTest extends BaseTestCase
 {
     public function testCreate()
     {